revise indexedStorageBuffer

+++ example/tech/initDrawIndexed.ts
@@ -4,7 +4,7 @@ const initDrawIndexed = (compiler: Compiler, colorAttachments: ColorAttachment[]
 
     let dispatch: RenderProperty;
     {
-        const indexData = new Int16Array([0, 1, 2, 0, 2, 3]);
+        const indexData = new Int32Array([0, 1, 2, 0, 2, 3]);
         const indexBuffer = compiler.createIndexBuffer({
             rawData: indexData
         });
